Move beachball config to its own package
