+++ packages/@rnw-scripts/beachball-config/.gitignore
@@ -1,0 +1,2 @@
+lib/
+lib-commonjs/
