+++ packages/@rnw-scripts/beachball-config/package.json
@@ -1,0 +1,41 @@
+{
+  "name": "@rnw-scripts/beachball-config",
+  "private": true,
+  "version": "0.0.0",
+  "license": "MIT",
+  "scripts": {
+    "build": "rnw-scripts build",
+    "clean": "rnw-scripts clean",
+    "lint": "rnw-scripts lint",
+    "lint:fix": "rnw-scripts lint:fix",
+    "watch": "rnw-scripts watch"
+  },
+  "main": "lib-commonjs/beachball.config.js",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/microsoft/react-native-windows.git",
+    "directory": "packages/@rnw-scripts/beachball-config"
+  },
+  "dependencies": {
+    "@react-native-windows/package-utils": "0.0.0-canary.24",
+    "@rnw-scripts/stamp-version": "0.0.0",
+    "find-up": "^4.1.0"
+  },
+  "devDependencies": {
+    "@rnw-scripts/eslint-config": "1.1.10",
+    "@rnw-scripts/just-task": "2.2.2",
+    "@rnw-scripts/ts-config": "2.0.1",
+    "@types/node": "^14.14.22",
+    "beachball": "^2.20.0",
+    "eslint": "^7.32.0",
+    "just-scripts": "^1.3.3",
+    "prettier": "^2.4.1",
+    "typescript": "^4.4.4"
+  },
+  "files": [
+    "lib-commonjs"
+  ],
+  "engines": {
+    "node": ">= 12.0.0"
+  }
+}
