+++ packages/@rnw-scripts/beachball-config/tsconfig.json
@@ -1,0 +1,5 @@
+{
+    "extends": "@rnw-scripts/ts-config",
+    "include": ["src"],
+    "exclude": ["node_modules"]
+}
